Move tagged_pair/tuple overloads for futures to algorithms module
