+++ libs/algorithms/include/hpx/parallel/algorithms/merge.hpp
@@ -12,7 +12,7 @@
 #include <hpx/concepts/concepts.hpp>
 #include <hpx/functional/invoke.hpp>
 #include <hpx/iterator_support/traits/is_iterator.hpp>
-#include <hpx/util/tagged_tuple.hpp>
+#include <hpx/parallel/util/tagged_tuple.hpp>
 
 #include <hpx/execution/algorithms/detail/is_negative.hpp>
 #include <hpx/execution/algorithms/detail/predicates.hpp>
