+++ libs/algorithms/include/hpx/parallel/algorithms/remove.hpp
@@ -13,8 +13,8 @@
 #include <hpx/concepts/concepts.hpp>
 #include <hpx/functional/invoke.hpp>
 #include <hpx/iterator_support/traits/is_iterator.hpp>
+#include <hpx/parallel/util/tagged_pair.hpp>
 #include <hpx/type_support/unused.hpp>
-#include <hpx/util/tagged_pair.hpp>
 
 #include <hpx/execution/algorithms/detail/is_negative.hpp>
 #include <hpx/execution/algorithms/detail/predicates.hpp>
