+++ libs/algorithms/include/hpx/parallel/algorithms/transform.hpp
@@ -19,10 +19,10 @@
 #include <hpx/functional/invoke.hpp>
 #include <hpx/functional/traits/is_callable.hpp>
 #include <hpx/iterator_support/traits/is_iterator.hpp>
+#include <hpx/parallel/util/tagged_pair.hpp>
+#include <hpx/parallel/util/tagged_tuple.hpp>
 #include <hpx/threading_base/annotated_function.hpp>
 #include <hpx/traits/segmented_iterator_traits.hpp>
-#include <hpx/util/tagged_pair.hpp>
-#include <hpx/util/tagged_tuple.hpp>
 
 #include <hpx/execution/execution_policy.hpp>
 #include <hpx/parallel/algorithms/detail/dispatch.hpp>
