+++ libs/algorithms/include/hpx/parallel/container_algorithms/minmax.hpp
@@ -15,7 +15,7 @@
 #include <hpx/concepts/concepts.hpp>
 #include <hpx/iterator_support/range.hpp>
 #include <hpx/iterator_support/traits/is_range.hpp>
-#include <hpx/util/tagged_pair.hpp>
+#include <hpx/parallel/util/tagged_pair.hpp>
 
 #include <hpx/parallel/algorithms/minmax.hpp>
 #include <hpx/parallel/traits/projected.hpp>
