+++ libs/algorithms/include/hpx/parallel/container_algorithms/partition.hpp
@@ -14,7 +14,7 @@
 #include <hpx/iterator_support/range.hpp>
 #include <hpx/iterator_support/traits/is_iterator.hpp>
 #include <hpx/iterator_support/traits/is_range.hpp>
-#include <hpx/util/tagged_tuple.hpp>
+#include <hpx/parallel/util/tagged_tuple.hpp>
 
 #include <hpx/parallel/algorithms/partition.hpp>
 #include <hpx/parallel/tagspec.hpp>
