+++ libs/algorithms/include/hpx/parallel/container_algorithms/remove.hpp
@@ -14,7 +14,7 @@
 #include <hpx/iterator_support/range.hpp>
 #include <hpx/iterator_support/traits/is_iterator.hpp>
 #include <hpx/iterator_support/traits/is_range.hpp>
-#include <hpx/util/tagged_pair.hpp>
+#include <hpx/parallel/util/tagged_pair.hpp>
 
 #include <hpx/parallel/algorithms/remove.hpp>
 #include <hpx/parallel/tagspec.hpp>
