+++ libs/algorithms/include_compatibility/hpx/util/tagged_tuple.hpp
@@ -1,0 +1,21 @@
+//  Copyright (c) 2020 STE||AR Group
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/config.hpp>
+#include <hpx/algorithms/config/defines.hpp>
+#include <hpx/parallel/util/tagged_tuple.hpp>
+
+#if defined(HPX_ALGORITHMS_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_MSVC)
+#pragma message("The header hpx/util/tagged_tuple.hpp is deprecated, \
+    please include hpx/datastructures/tagged_tuple.hpp or \
+    hpx/parallel/util/tagged_tuple.hpp instead")
+#else
+#warning "The header hpx/util/tagged_tuple.hpp is deprecated, \
+    please include hpx/datastructures/tagged_tuple.hpp or \
+    hpx/parallel/util/tagged_tuple.hpp instead"
+#endif
+#endif
